Back_End/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
